Refactors, better logs, verbose & overwrite flag, documentation
