+++ src/discord.rs
@@ -1,18 +1,19 @@
 use crate::{
     config::{read_config_file, write_config, Config, DiscordConfig},
     parser::CliDiscordSet,
-    processes::{get_processes, replace_asset},
+    processes::{get_data, get_names},
 };
 use discord_rich_presence::{activity::*, DiscordIpc, DiscordIpcClient};
-// use discord_rpc_client::{models::ActivityAssets, Client};
 use std::{collections::HashMap, io::Error, process::exit};
-use tracing::{error, info, instrument, trace, warn};
+use tracing::{debug, error, info, instrument, trace, warn};
 
+/// Bundle DiscordIpcClient and the Discord activity data associated to it.
 pub struct DiscordClientWrapper {
     client: DiscordIpcClient,
     replaced_data: DiscordConfig,
 }
 
+/// Print Discord activity data saved in config.
 pub fn get_activity_data(config: &DiscordConfig) -> () {
     println!(
         "Client ID: {}\nDetails: {}\nState: {}\nLarge Image Key: {}\nLarge Image Text: {}\nSmall Image Key: {}\nSmall Image Text: {}\nButton 1 Text: {}\nButton 1 URL: {}\nButton 2 Text: {}\nButton 2 URL: {}",
@@ -30,7 +31,11 @@ pub fn get_activity_data(config: &DiscordConfig) -> () {
     );
 }
 
+/// Overwrite Discord data in `Config` and write to file.
+#[instrument(skip_all)]
 pub fn set_activity_data(config: &mut Config, arg: CliDiscordSet) -> () {
+    trace!("Overwriting with:\n{arg:#?}");
+
     if let Some(id) = arg.client_id {
         config.discord.client_id = id
     }
@@ -72,16 +77,26 @@ pub fn set_activity_data(config: &mut Config, arg: CliDiscordSet) -> () {
     write_config(&config);
 }
 
+/// Initialize and connect `DiscordIpcClient`.
 #[instrument(skip_all)]
 pub fn client_init(client_id: u64) -> DiscordClientWrapper {
-    let mut client = DiscordIpcClient::new(&client_id.to_string()).unwrap();
+    let mut client: DiscordIpcClient = match DiscordIpcClient::new(&client_id.to_string()) {
+        Err(error) => {
+            error!("Unable to initialize Discord client: {error}");
+            exit(1);
+        }
+        Ok(client) => {
+            trace!("Successfully initialized Discord client");
+            client
+        }
+    };
 
     match client.connect() {
         Err(err) => {
-            error!("{err}");
-            exit(1)
+            error!("Error while connect Discord client to IPC: {err}");
+            exit(1);
         }
-        Ok(_) => info!("Discord Client connected to IPC"),
+        Ok(_) => info!("Discord client connected to IPC"),
     }
 
     return DiscordClientWrapper {
@@ -90,49 +105,59 @@ pub fn client_init(client_id: u64) -> DiscordClientWrapper {
     };
 }
 
+/// Create the hashmap for template variables and their replacements in Discord data.
 #[instrument(skip_all)]
 fn template_hashmap(config: &Config) -> HashMap<&str, String> {
-    let processes = get_processes(&config.processes);
+    let processes = get_names(&config.processes);
 
     let mut replace_hashmap: HashMap<&str, String> = HashMap::new();
-    replace_hashmap.insert(
-        "process.text",
-        replace_asset(&config.processes, &processes).0,
-    );
-    replace_hashmap.insert(
-        "process.icon",
-        replace_asset(&config.processes, &processes).1,
-    );
+    replace_hashmap.insert("process.icon", get_data(&config.processes, &processes).1);
+    replace_hashmap.insert("process.text", get_data(&config.processes, &processes).0);
+    replace_hashmap.insert("idle.icon", config.processes.idle_icon.to_owned());
+    replace_hashmap.insert("idle.text", config.processes.idle_text.to_owned());
 
+    trace!("Template variable hashmap created");
     return replace_hashmap;
 }
 
+/// Replace recognized template variables with their corresponding data.
 #[instrument(skip_all)]
-pub fn replace_variables(template_hashmap: &HashMap<&str, String>, mut string: String) -> String {
-    if string.is_empty() {
+pub fn replace_template_variables(
+    template_hashmap: &HashMap<&str, String>,
+    mut string: String,
+) -> String {
+    if string.is_empty() || !string.contains("{{") {
+        trace!("String does not contain template variables");
         return string;
     }
     for (target, replacement) in template_hashmap {
         let target = format!("{{{{{}}}}}", target);
+        // trace!("Replacing template variable \"{target}\" with \"{replacement}\"");
         string = string.replace(&target, replacement)
     }
 
     return string;
 }
 
+/// Set Discord activity. Will clone `DiscordConfig` data and replace template variables before comparing to old data. If the new data matches<br/>
+/// with the old data, the function will return. Otherwise, the new data is used and the activity will be overwritten.
 #[instrument(skip_all)]
 pub fn set_activity(
     mut client_wrapper: DiscordClientWrapper,
     config: &mut Config,
 ) -> DiscordClientWrapper {
     let mut replaced_data: DiscordConfig = config.discord.clone();
+    trace!("Discord data cloned");
 
     let template_hashmap: HashMap<&str, String> = template_hashmap(config);
     replaced_data = replaced_data.replace_templates(&template_hashmap);
 
     if replaced_data == client_wrapper.replaced_data {
+        debug!("Activity data has not changed");
         return client_wrapper;
     }
+
+    trace!("Activity data has changed, overwriting and setting activity");
 
     client_wrapper.replaced_data = replaced_data;
 
@@ -195,18 +220,26 @@ pub fn set_activity(
     return client_wrapper;
 }
 
+/// Clears the current Discord activity
 #[instrument(skip_all)]
-pub fn clear_activity(mut client: DiscordIpcClient) -> Result<DiscordIpcClient, Error> {
-    client.clear_activity().unwrap();
-    info!("Discord RPC activity cleared");
-    return Ok(client);
+pub fn clear_activity(mut wrapper: DiscordClientWrapper) -> Result<DiscordClientWrapper, Error> {
+    wrapper.client.clear_activity().unwrap();
+    info!("Discord activity cleared");
+    return Ok(wrapper);
 }
 
+/// Updates `Config` and sets Discord activity if no errors occur during config reread. If an error does occur, a warning will be logged<br/>
+/// but no changes will take place.
 #[instrument(skip_all)]
 pub fn update_activity(config: &mut Config, client: DiscordClientWrapper) -> DiscordClientWrapper {
-    *config = read_config_file();
-    info!("Updating Discord RPC");
+    *config = match read_config_file(true) {
+        Err(_) => {
+            warn!("Config file was not deserialized. Will continue to use old config.");
+            return client;
+        }
+        Ok(config) => config,
+    };
+    info!("Updating Discord activity");
     let client: DiscordClientWrapper = set_activity(client, config);
-    info!("Updated Discord RPC");
     return client;
 }
