--- src/parser/mod.rs
@@ -1,4 +1,0 @@
-pub mod cli;
-mod structure;
-
-pub use structure::*;
